Remove rustfmt.toml
The specified attribute is only available with a nightly compiler.
I don't typically use a nightly compiler, and none of our CI jobs
are set up with a nightly compiler. With this configuration in place,
the code generation job generates many unnecessary warnings.

If we'd like to maintain this configuration, we should

- update our Makefile to specify a nightly compiler
- update our CI jobs to use a nightly compiler