Generate const generic instances
Before this commit, all peripheral instances shared the same instance
type, Instance. This commit introduces a const generic N to indicate the
peripheral's instance number. The update allows users to design APIs
that are generic over all N, or specific to an instance number. It's a
refactor of a previous approach that used typenums.

Instance APIs are now behind an impl block, instead of a dedicated
module. It simplifies the auto-generated code, and makes it easier for
users to design APIs to the instance types. It's also backwards
compatible with the read_reg!, write_reg!, and modify_reg! macros.
Unfortunately, it breaks reset_reg!, since the macro assumed a path
instead of an impl block. A child commit updates the macro for the new
implementation.

A "sole instance" is a peripheral that has only one instance. We use 0
as the magic const generic. This isn't a detail users will need to
remember; the usage guide (incoming child commit) shows how users should
design APIs for a sole instance.

Commit builds with all chip features. Also build the 1062 with 'nosync'
and 'rtic' features. Tests fail to compile; that's fixed in child
commits.