Remove nosync examples
I'm removing support for raw pointer access in ral-registers 0.2. The
API forms a reference to MMIO memory, which is considered a no-no. And
in my experience, most folks don't know that this was an API feature.

Folks who want something like this are free to call the various instance
methods. These already require an unsafe block.